service layer fix
